refactor(dao): move common abstract/flow code to a base DAO
